shopping-backend/requirements.txt
psycopg2
django
django-cors-headers
django-rest-knox
knox
djangorestframework